+++ src/core/interaction/pickingmanager.cpp
@@ -157,9 +157,9 @@ bool PickingManager::isPickingActionRegistered(const PickingAction* action) cons
 // Then all adjacent pairs are swapped and then all adjacent single bits.
 // This results in a reversed order.
 std::uint8_t reverse(std::uint8_t b) {
-    b = (b & 0xF0) >> 4 | (b & 0x0F) << 4;
-    b = (b & 0xCC) >> 2 | (b & 0x33) << 2;
-    b = (b & 0xAA) >> 1 | (b & 0x55) << 1;
+    b = static_cast<std::uint8_t>((b & 0xF0) >> 4 | (b & 0x0F) << 4);
+    b = static_cast<std::uint8_t>((b & 0xCC) >> 2 | (b & 0x33) << 2);
+    b = static_cast<std::uint8_t>((b & 0xAA) >> 1 | (b & 0x55) << 1);
     return b;
 }
